# openapi.yaml
openapi: 3.0.3
info:
  title: ReviewDunia API
  version: "0.1.0"
  description: API for reviewdunia.com - product catalog, uploads, review generation, admin approval.
servers:
  - url: https://api.reviewdunia.local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        brand:
          type: string
        category_id:
          type: string
        primary_image_path:
          type: string
        affiliate_links:
          type: object
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/ProductSummary'
        - type: object
          properties:
            description:
              type: string
            status:
              type: string
            latest_review:
              type: object
...
paths:
...
  /api/v1/products:
    get:
      summary: Search products by text
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductSummary'
  /api/v1/products/{product_id}:
    get:
      summary: Get product detail
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '404':
          description: Not found
...